chore/fix: extract helper fns to designated files, use T_DIR for direct encoding
